XOR works properly, running into issues with Spring Boot and Java.lang.String beans? Need to learn more about Spring
